refactor: store, zustand 디렉토리 분리
