+++ apps/frontend/src/_storybook/layout/Header/HeaderUserInfo.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import HeaderUserInfo from "@/_components/layout/Header/HeaderUserInfo";
-import useUserStore from "@/_store/useUserStore";
+import useUserStore from "@/_store/zustand/useUserStore";
 
 const mockUser = {
   id: 1,
