+++ apps/frontend/src/_storybook/pages/chat/ChatSection.stories.tsx
@@ -6,8 +6,8 @@ import {
 } from "@/_store/redux/type";
 import { Provider } from "react-redux";
 import { configureStore } from "@reduxjs/toolkit";
-import { useInterviewerStore } from "@/_store/useInterviewerStore";
-import useUserStore from "@/_store/useUserStore";
+import { useInterviewerStore } from "@/_store/zustand/useInterviewerStore";
+import useUserStore from "@/_store/zustand/useUserStore";
 import chatReducer from "@/_store/redux/features/chat/slice";
 
 const mockInterviewer = {
